Separating routes into a different file for better organization.
